RPGLibrary/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.2
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
